+++ Back/Task_Manager_Back/Task_Manager_Back.Application/Requests/TaskRequests/UpdateTaskRequest.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Task_Manager_Back.Domain.Entities.Enums;
+using Task_Manager_Back.Domain.Entities.TaskRelated;
 
 namespace Task_Manager_Back.Application.Requests.TaskRequests;
 public record UpdateTaskRequest(
@@ -9,6 +10,7 @@ public record UpdateTaskRequest(
     string NewDescription,
     Guid NewStatusId,
     Guid NewCategoryId,
+    TaskLocation NewLocation,
     DateTime NewDeadline,
     TaskPriority NewPriority,
     string NewColor,
@@ -16,5 +18,5 @@ public record UpdateTaskRequest(
     int Price,
     int NewPositionOrder,
     int NewPrice,
-    List<Guid> NewDependsOnTasksIds // NEW
+    List<Guid> NewDependsOnTasksIds
 ) : IRequest;